first screen is added

+++ lib/movies/data/models/movies_model.dart
@@ -19,8 +19,7 @@ class MoviesModel extends Movie{
     overview: json['overview'], 
     backdropPath: json['backdrop_path'], 
     generId:List<int>.from(json['genre_ids'].map((e) => e)), 
-    // TODO check this 
-    voteAverage: json['vote_average'], 
+    voteAverage: json['vote_average'].toDouble(), 
     releaseDate: json['release_date']
     );
